[LiveSimplification] Remove the Peaking Lot Preview
